requirements.txt
GDAL==3.0.4
numpy
matplotlib
rasterio
shapely
geopandas
earthpy
Rtree
earthpy
seaborn
